computer-control-agent/pyproject.toml
[project]
name = "class13-14"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.14",
    "playwright>=1.52.0",
]
